Add Dij's Snow Wars + ??next move processing??
I wrapped in Dij's Snow Wars script, and then I added in a feature where it does what a human would do and finishes attacking the 2x2, 2x1, and 1x3 structures uncovered when proceeding in Dij's grid pattern, returning to that pattern once the structure is obliterated.  I don't know if this would be unacceptable to GC devs since it does the same thing a human would do, but I'm not sure!  Anyway, I am not using it on the site until I figure out how to get it approved somehow...